컴파일 가능_1105

+++ src/userprog/syscall.h
@@ -1,6 +1,7 @@
 #ifndef USERPROG_SYSCALL_H
 #define USERPROG_SYSCALL_H
 #include <stdbool.h>
+#include "filesys/off_t.h"
 
 void syscall_init (void);
 
@@ -15,8 +16,8 @@ int syscall_open(const char *file);
 int syscall_filesize (int fd);
 int syscall_read (int fd, void *buffer, unsigned size);
 int syscall_write(int fd, void *buffer, unsigned size);
-void syscall_seek (int fd, unsigned position);
-unsigned syscall_tell (int fd);
+void syscall_seek (int fd, off_t position);
+off_t syscall_tell (int fd);
 void syscall_close (int fd);
 
 #endif /* userprog/syscall.h */
